HouseWithMedianPrice.java
You are given information of N houses, each with price, number of bedrooms, and square footage and asked to output the information for a house that has the median price among all houses.

The median is the middle number in a sorted, ascending or descending, list of numbers. N is guanranteed to be odd.

Input Format

The first line contains one integer number N, the number of houses.

Each of the next N lines contains 3 integer numbers which are price, number of bedrooms, and square footage.

Constraints

1 <= N <= 10^6 N is odd.

0 < price <= 1000000 and all prices will be distinct.

0 < number_of_bedrooms <= 10

0 < square_footage <= 10000

Output Format

Output one line, the house that has median price.